refactor: business-hours.ts -> business-hours-types.ts 잘못된 파일명 수정
